requirements.txt
openenv-core>=0.2.2
fastapi
uvicorn
pydantic
openai
huggingface-hub